Automation for the Login module.
