+++ package.json
@@ -1,0 +1,18 @@
+{
+  "name": "hrportal_playwright",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {},
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "devDependencies": {
+    "@playwright/test": "^1.48.0",
+    "@types/node": "^22.7.5"
+  },
+  "dependencies": {
+    "mocha": "^11.0.1",
+    "mochawesome": "^7.1.3"
+  }
+}
